+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Bugfixes:
 - Don't block playback for CoreAudio/Wasapi if there is no data in time.
 - Validate `silence_threshold` and `silence_timeout` fields.
 - Fix panic when reloading config if a new filter was defined but not added to the pipeline.
+- Check for mixer parameter changes when reloading config.
+ 
 
 
 ## 0.4.0